Move ContinerEnvVar class to the template subpackage
