+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/ContainerUtilsTest.java
@@ -13,8 +13,8 @@ import io.fabric8.kubernetes.api.model.LifecycleBuilder;
 import io.fabric8.kubernetes.api.model.Quantity;
 import io.fabric8.kubernetes.api.model.ResourceRequirementsBuilder;
 import io.fabric8.kubernetes.api.model.SecurityContextBuilder;
-import io.strimzi.api.kafka.model.common.ContainerEnvVarBuilder;
 import io.strimzi.api.kafka.model.common.Probe;
+import io.strimzi.api.kafka.model.common.template.ContainerEnvVarBuilder;
 import io.strimzi.api.kafka.model.common.template.ContainerTemplate;
 import io.strimzi.api.kafka.model.common.template.ContainerTemplateBuilder;
 import io.strimzi.operator.common.Reconciliation;
