+++ packaging/helm-charts/helm3/strimzi-kafka-operator/crds/040-Crd-kafka.yaml
@@ -1835,7 +1835,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -1978,7 +1978,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -3261,7 +3261,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -4503,7 +4503,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -4646,7 +4646,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -4789,7 +4789,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -5957,7 +5957,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -6100,7 +6100,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -7067,7 +7067,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
@@ -7963,7 +7963,7 @@ spec:
                                           configMapKeyRef: {}
                                         required:
                                           - configMapKeyRef
-                                    description: Reference to secret or config map property to which will be the environment variable is set.
+                                    description: Reference to the secret or config map property to which the environment variable is set.
                                 oneOf:
                                   - properties:
                                       value: {}
