+++ packaging/helm-charts/helm3/strimzi-kafka-operator/crds/041-Crd-kafkaconnect.yaml
@@ -1223,7 +1223,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
@@ -1366,7 +1366,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
@@ -2150,7 +2150,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
