+++ packaging/helm-charts/helm3/strimzi-kafka-operator/crds/045-Crd-kafkamirrormaker.yaml
@@ -1344,7 +1344,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
