+++ packaging/helm-charts/helm3/strimzi-kafka-operator/crds/046-Crd-kafkabridge.yaml
@@ -1201,7 +1201,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
@@ -1380,7 +1380,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
