+++ packaging/helm-charts/helm3/strimzi-kafka-operator/crds/048-Crd-kafkamirrormaker2.yaml
@@ -1440,7 +1440,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
@@ -1583,7 +1583,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
@@ -2367,7 +2367,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                       - configMapKeyRef
-                                description: Reference to secret or config map property to which will be the environment variable is set.
+                                description: Reference to the secret or config map property to which the environment variable is set.
                             oneOf:
                               - properties:
                                   value: {}
