+++ packaging/install/cluster-operator/040-Crd-kafka.yaml
@@ -1834,7 +1834,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -1977,7 +1977,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -3260,7 +3260,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -4502,7 +4502,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -4645,7 +4645,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -4788,7 +4788,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -5956,7 +5956,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -6099,7 +6099,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -7066,7 +7066,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
@@ -7962,7 +7962,7 @@ spec:
                                       configMapKeyRef: {}
                                     required:
                                     - configMapKeyRef
-                                  description: Reference to secret or config map property to which will be the environment variable is set.
+                                  description: Reference to the secret or config map property to which the environment variable is set.
                               oneOf:
                               - properties:
                                   value: {}
