+++ packaging/install/cluster-operator/046-Crd-kafkabridge.yaml
@@ -1200,7 +1200,7 @@ spec:
                                   configMapKeyRef: {}
                                 required:
                                 - configMapKeyRef
-                              description: Reference to secret or config map property to which will be the environment variable is set.
+                              description: Reference to the secret or config map property to which the environment variable is set.
                           oneOf:
                           - properties:
                               value: {}
@@ -1379,7 +1379,7 @@ spec:
                                   configMapKeyRef: {}
                                 required:
                                 - configMapKeyRef
-                              description: Reference to secret or config map property to which will be the environment variable is set.
+                              description: Reference to the secret or config map property to which the environment variable is set.
                           oneOf:
                           - properties:
                               value: {}
