+++ packaging/install/cluster-operator/048-Crd-kafkamirrormaker2.yaml
@@ -1439,7 +1439,7 @@ spec:
                                   configMapKeyRef: {}
                                 required:
                                 - configMapKeyRef
-                              description: Reference to secret or config map property to which will be the environment variable is set.
+                              description: Reference to the secret or config map property to which the environment variable is set.
                           oneOf:
                           - properties:
                               value: {}
@@ -1582,7 +1582,7 @@ spec:
                                   configMapKeyRef: {}
                                 required:
                                 - configMapKeyRef
-                              description: Reference to secret or config map property to which will be the environment variable is set.
+                              description: Reference to the secret or config map property to which the environment variable is set.
                           oneOf:
                           - properties:
                               value: {}
@@ -2366,7 +2366,7 @@ spec:
                                   configMapKeyRef: {}
                                 required:
                                 - configMapKeyRef
-                              description: Reference to secret or config map property to which will be the environment variable is set.
+                              description: Reference to the secret or config map property to which the environment variable is set.
                           oneOf:
                           - properties:
                               value: {}
