+++ packaging/install/cluster-operator/04A-Crd-kafkanodepool.yaml
@@ -955,7 +955,7 @@ spec:
                                   configMapKeyRef: {}
                                 required:
                                 - configMapKeyRef
-                              description: Reference to secret or config map property to which will be the environment variable is set.
+                              description: Reference to the secret or config map property to which the environment variable is set.
                           oneOf:
                           - properties:
                               value: {}
@@ -1098,7 +1098,7 @@ spec:
                                   configMapKeyRef: {}
                                 required:
                                 - configMapKeyRef
-                              description: Reference to secret or config map property to which will be the environment variable is set.
+                              description: Reference to the secret or config map property to which the environment variable is set.
                           oneOf:
                           - properties:
                               value: {}
